LevelGeneratorManagerV2
DO NOT USE THE OLD ONE, it's been deprecated so you shouldnt be able to use it anyway.

No more BP_Hallway / BP_StartRoom.
Place the Generator and any number of RoomSpawnLocations into a level and rooms will be spawned automatically.
Per-Location overrides can be specified to change what sort of rooms spawn where. Previews are shown to make it easy to visually distinguish different spawn types.
Actual generation code is simpler and the logic rewritten to be more performant.