Setup Data Stuff for Spawning
Updated Data Table so every room type has an entry for spawning.
Updated defaults in level generator so that each override type spawns the correct rooms and none are left empty, so no erroring.

Added placeholders for each room type so that they exist for playtesting. Will be updated later on